CourseProject
The goal of this project is to predict the manner in which people did the exercise. This is the "classe" variable in the training set. I used 58 variables as predictors to predict the "classe" output. I used the prediction model to predict 20 different test cases of Quiz 4, and I got 100% correct prediction.